benchmarks/poco_cppzmq_2/client/requirements.txt
numpy==1.21.0
pyzmq==22.3.0
